refactor(tinacms): extract media cmpnts to own files
