+++ packages/tinacms/src/components/media/index.ts
@@ -17,3 +17,6 @@ limitations under the License.
 */
 
 export * from './media-manager'
+export { MediaItem } from './media-item'
+export { Breadcrumb } from './breadcrumb'
+export { PageLinks } from './pagination'
